# src/main/resources/AwsServicesOpenApi.yaml
openapi: 3.0.1
info:
  title: AwsServicesTest
  description: This APi test or implements diffrent AWS services
  termsOfService: NA
  contact:
    email: er.vikaspaul@gmail.com
  license:
    name: AwsServices
  version: 1.0.0
...
servers:
  - url: 'https://localhost:7070/awsservices/v1'
...
paths:
...
  /user:
    post:
      tags:
        - user
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '202':
          description: User Sucessfully Created
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '405':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        default:
          description: successful operation
          content: {}
    put:
      tags:
        - user
      summary: Updated user
      description: This can only be done by the logged in user.
      operationId: updateUser
      parameters:
        - name: username
          in: path
          description: name that need to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Updated user object
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '202':
          description: User Sucessfully updated
          content: {}
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '405':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        default:
          description: successful operation
          content: {}
  /user/createWithArray:
    post:
      tags:
        - user
      summary: Creates list of users with given input array
      description: Creates list of users with given input array
      operationId: createUsersWithArrayInput
      requestBody:
        description: List of user object
        content:
          '*/*':
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        '202':
          description: User Creted With List Sucessful
          content: {}
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '405':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        default:
          description: successful operation
          content: {}
  /user/login:
    get:
      tags:
        - user
      summary: Logs user into the system
      description: Logs user into the system
      operationId: loginUser
      parameters:
        - name: username
          in: query
          description: The user name for login
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: The password for login in clear text
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User Sucessfully Loggedin
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Expires-After:
              description: date in UTC when token expires
              schema:
                type: string
                format: date-time
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '405':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        default:
          description: successful operation
          content: {}
  /user/logout:
    get:
      tags:
        - user
      summary: Logs out current logged in user session
      operationId: logoutUser
      responses:
        '202':
          description: User Logout Sucessful
          content: {}
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '405':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        default:
          description: successful operation
          content: {}
  '/user/{username}':
    get:
      tags:
        - user
      summary: Get user by user name
      operationId: getUserByName
      parameters:
        - name: username
          in: path
          description: 'The name that needs to be fetched. Use user1 for testing. '
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User Details Sucessfully retrived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Userdetails'
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '405':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        default:
          description: successful operation
          content: {}
    delete:
      tags:
        - user
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      parameters:
        - name: username
          in: path
          description: The name that needs to be deleted
          required: true
          schema:
            type: string
      responses:
        '202':
          description: User Sucessfully Deleted
          content: {}
        '400':
          description: Invalid ID supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '405':
          description: Validation exception
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        default:
          description: successful operation
          content: {}
components:
  schemas:
...
    CreateUserRequest:
      type: object
      required: 
        - customerUUID
        - contactName
        - contactEmail
        - contactNumber
        - type
        - role
      properties:
        customerUUID:
          description: CustomerID of the Registering Customer
          items:
            $ref: '#/components/schemas/CustomerUUID'
        contactName:
          description: Contact person of thest company
          items:
            $ref: '#/components/schemas/UserContactName'
        contactEmail:
          description: Company Contact Email
          items:
            $ref: '#/components/schemas/UserContactEmail'
        contactNumber:
          description: Contact number of the customer
          items:
            $ref: '#/components/schemas/UserContactNumber'
        type:
          description: Type of Customer
          items:
            $ref: '#/components/schemas/UserType'
        role:
          description: USer Role
          items:
            $ref: '#/components/schemas/UserRole'
    Userdetails:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerUUID:
          description: CustomerID of the Registering Customer
          items:
            $ref: '#/components/schemas/CustomerUUID'
        ContactName:
          description: Contact person of thest company
          items:
            $ref: '#/components/schemas/UserContactName'
        ContactEmail:
          description: Company Contact Email
          items:
            $ref: '#/components/schemas/UserContactEmail'
        ContactNumber:
          description: Contact number of the customer
          items:
            $ref: '#/components/schemas/UserContactNumber'
        Type:
          description: Type of Customer
          items:
            $ref: '#/components/schemas/UserType'
        Role:
          description: USer Role
          items:
            $ref: '#/components/schemas/UserRole'
    ApiErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
...
    CustomerUUID:
      type: string
    UserContactName:
      type: string
      minLength: 1
      maxLength: 50
    UserContactEmail:
      type: string
    UserContactNumber:
      type: integer
      format: int32
    UserType:
      type: string
      enum:
        - ADMIN
        - SUBADMIN
        - USER
        - DUMMY
    UserRole:
      type: string
      enum:
        - ADMIN
        - SUBADMIN
        - USER
        - DUMMY